% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntax_composer.R
\name{.make_loadings_dt}
\alias{.make_loadings_dt}
\title{Make loadings DT
Given a psych::efa object, make a loadings dt for syntax_composer}
\usage{
.make_loadings_dt(fa_object, factor_names)
}
\arguments{
\item{factor_names}{}
}
\value{
long_loadings data.table
}
\description{
Make loadings DT
Given a psych::efa object, make a loadings dt for syntax_composer
}
