% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esemComp-package.R
\docType{package}
\name{esemComp-package}
\alias{esemComp}
\alias{esemComp-package}
\title{esemComp: ESEM-within-CFA syntax composer}
\description{
Functions to compose ESEM-within-CFA models to run in lavaan.
}
\author{
\strong{Maintainer}: Mateus Silvestrin \email{mateusilvestrin@gmail.com} (\href{https://orcid.org/0000-0002-3482-3676}{ORCID})

Authors:
\itemize{
  \item Leon T. de Beer \email{abcd@fgh.com} (\href{https://orcid.org/0000-0001-6900-2192}{ORCID})
}

}
\keyword{internal}
