% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra.R
\name{export_lavaan_results}
\alias{export_lavaan_results}
\title{Export (dump) lavaan results}
\usage{
export_lavaan_results(
  lavaan_model_fit,
  path_name = "lavaan_summary.txt",
  preamble = NULL,
  ...
)
}
\arguments{
\item{lavaan_model_fit}{Lavaan object. Lavaan object with fitted ESEM model, usually obtained with \code{lavaan::cfa()}.}

\item{path_name}{Character vector. Path directing where to save the text file with the results. It must end with the
desired file name with the desired text file extension. The default is to save the file in the current directory
with file name "lavaan_summary.txt".}

\item{preamble}{Character vector. Allows the user to add some text before the results report. Useful to
identify the analysis, add information about the data etc.}

\item{...}{\code{parameter = value} pairs. Additional modifiers to lavaan::summary() call. See the full list
of available modifiers at \cite(lavaan-class).}
}
\value{
Message confirming data dump. Saves the data to a text file as a side-effect.
}
\description{
Export (dump) lavaan results
}
\details{
Runs \code{lavaan::summary()} on \code{lavaan_model_fit} with modifiers \code{std = TRUE}
and \code{fit.measures = TRUE} and dumps the result to a text file. The user has the option to add some
text as a preamble before the results description. It is also possible to add more information to the
exported results by including more modifiers to the \code{summary} call with the \code{...} named values
(e.g. \code{rsquare = TRUE} to include r-squared measures), see \cite(lavaan-class) for the full list
of available modifiers.

The \code{path_name} directories must be separated by "\" (even in Windows).
}
\examples{
HS.model <- ' visual  =~ x1 + x2 + x3
textual =~ x4 + x5 + x6
speed   =~ x7 + x8 + x9 '
fit <- lavaan::cfa(HS.model, data= lavaan::HolzingerSwineford1939)

export_lavaan_results(fit)
# name the analysis with preamble
export_lavaan_results(fit,
path_name = "titled_lavaan_summary.txt",
preamble = "CLASSIC HOLZINGER SWINEFORD CFA MODEL")
# add r-squared measure to the exported results
export_lavaan_results(fit,
path = "lavaan_summary_rsquared.txt",
rsquare = TRUE)
}
