% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntax_composer.R
\name{find_referents}
\alias{find_referents}
\title{Automatically find referent items for ESEM}
\usage{
find_referents(efa_object, factor_names)
}
\arguments{
\item{efa_object}{A \cite{psych::fa()} object with the results of a exploratory factor analysis.
Referents will be selected based on the factor loadings in this solution.}

\item{factor_names}{Character vector. The names should identify the factors in the \code{efa_object}.
They must be in the same order they appear in the \code{efa_object}.}
}
\value{
A named list in the format \code{c(FactorName = referent)}, in the same order as the given
factor names.
}
\description{
Automatically find referent items for ESEM
}
\details{
Given an exploratory factor analysis results, this function finds a referent item for each
factor and outputs a list to be used in \cite{syntax_composer}.

The automatic selection currently does not choose the referents with the recommended approach
in the literature, that is, items that load heavily on one factor and lightly on all the others.
In its current implementation, the items are chosen solely based on their highest loadings.
This may lead to less than ideal referent selection in some situations. It is recommended to always
compare the resulting  referents with the items one would choose when inspecting the exploratory
solution loadings (usually with \code{loadings(efa_object)}). In fact, the user should always check
\code{loadings(efa_object)} \emph{before} calling the function to be able to specify \code{factor_names} correctly.
}
\examples{
# use Holzinger and Swineford (1939) dataset in lavaan package
hw_data <- lavaan::HolzingerSwineford1939
hw_data <- hw_data[,c(7:15)]

#make exploratory analysis with geomin rotation
geomin_efa <- esem_efa(hw_data,3)

# check the order of factors in the efa solution
loadings(geomin_efa)
#find referents with factors in the order checked above
find_referents(geomin_efa, c("textual", "visual", "speed"))
#In this particular case, automatic selection chooses
#the same items one would choose manually. For comparison:
loadings(geomin_efa)
}
