% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra.R
\name{fit_free_factor_var_esem}
\alias{fit_free_factor_var_esem}
\title{Model fitting for free factor variance ESEM}
\usage{
fit_free_factor_var_esem(model_syntax, data, ...)
}
\arguments{
\item{model_syntax}{An ESEM model fitting create with \cite{syntax_composer()}
and \code{only_fix_crossloadings = FALSE}.}

\item{data}{Data.frame where rows are observations and columns are indicators. Alternatively,
a covariance matrix. If a covariance matrix, \code{sample.nobs} must be included in \code{...}}

\item{...}{Other \code{name = value} pairs to \cite{lavaan::lavaan()}.}
}
\value{
A lavaan object with the fitted model.
}
\description{
Model fitting for free factor variance ESEM
}
\details{
Lavaan's \cite{lavaan::cfa()} function automatically fixes the first indicator of
factors when factor (residual) variances are set as free parameters with \code{std.lv = TRUE}.
The present function is a wrapper around \cite{lavaan::lavaan()} with the same parameters as
the cfa function, except that factor variances are set free and no additional fixing is done.
This allows the correct estimation of ESEM-within-CFA models with free factor variances, granted
that the model syntax was made with \cite{syntax_composer()} with \code{only_fix_crossloadings = FALSE} and
thus includes the correct number of fixed parameters. See the \cite{esem-as-efa} vignette for an example.
}
