% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_target.R
\name{make_target}
\alias{make_target}
\title{Make target}
\usage{
make_target(nitems, mainloadings, bifactor = FALSE)
}
\arguments{
\item{nitems}{An integer. The total number of items.}

\item{mainloadings}{A list. A list indicating the indexes of the items related to each latent variable. See examples.}

\item{bifactor}{Logical. If TRUE, adds a G-factor column to the resulting target matrix.}
}
\value{
Target matrix with NAs and zeros [nItems x nLatentVariables].
}
\description{
Make target rotation matrix for use with \cite{esem_efa()}.
}
\details{
A target rotation matrix is composed of cells indicating which loadings should be freely estimated
and which should be as close to zero as possible when the factor solution is rotated. Freely
estimated loadings are represented with NAs. \code{mainloadings} identify which items are free
for each latent variable. See the vignette esem as efa for suggestions on how to use.
}
\examples{
#target matrix for the Tucker dataset in the psych package
make_target(9, list(f1 = c(1,2,5:7), f2 = c(3,4,8,9)))

#bifactor matrix for the same dataset
make_target(9, list(f1 = c(1,2,5:7), f2 = c(3,4,8,9)),TRUE)
}
