% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntax_composer.R
\name{syntax_composer}
\alias{syntax_composer}
\title{ESEM-within-CFA syntax composer for lavaan}
\usage{
syntax_composer(efa_object, referents, only_fix_crossloadings = TRUE)
}
\arguments{
\item{efa_object}{A \cite{psych::fa()} object with the results of a exploratory factor analysis.
The factor loadings from this analysis will be set as starting values or fixed values in the
lavaan syntax created.}

\item{referents}{A named list in the form \code{list(factor1 = "referent1", etc)}. Each
entry identifies the referent for one of the factors. The order of this vector must be
the same order the factors are ordered in the \code{efa_object}.}
}
\value{
A character vector with a lavaan syntax for the ESEM model.
}
\description{
ESEM-within-CFA syntax composer for lavaan
}
\details{
ESEM-within-CFA models have latent variables estimations with all cross-loadings.
To improve fit, starting values from a previous exploratory analysis are used. To ensure
model identification, a referent is chosen for each latent variable and all their
cross-loadings are set as fixed values. This function "composes" such a lavaan model
syntax from an EFA object. This model can subsequently be used in lavaan with
\cite{lavaan::cfa()} or \cite{lavaan::sem()}. The best way to check the output is
with \cite{writeLines()}.

It is very important to make sure the order of the \code{referents} vector is the same as the
factor order in the factor loadings matrix in \code{efa_object}. Otherwise, the names of the
latent variables may be wrong in the syntax. See also the vignette \cite{esem-within-cfa}.
}
\examples{
# use Holzinger and Swineford (1939) dataset in lavaan package
hw_data <- lavaan::HolzingerSwineford1939
hw_data <- hw_data[,c(7:15)]

#make exploratory analysis with geomin rotation
geomin_efa <- esem_efa(hw_data,3)
referents_list <- list(textual = "x5", visual = "x3", speed = "x7")
model_syntax <- syntax_composer(geomin_efa, referents_list)
writeLines(model_syntax)

# esem-within-cfa
esem_w_cfa <- lavaan::cfa(model_syntax, data = hw_data, std.lv = TRUE)
summary(esem_w_cfa)
}
\seealso{
\cite{make_target}
}
