% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_target.R
\name{esem_efa}
\alias{esem_efa}
\title{ESEM EFA}
\usage{
esem_efa(
  data,
  nfactors,
  target,
  bifactor = FALSE,
  fm = "pa",
  rotate = "targetQ",
  ...
)
}
\arguments{
\item{data}{Raw Data.frame or matrix. [Subjects x Items] data or covariance matrix.}

\item{nfactors}{An integer. Number of factors to extract (including G when bifactor).}

\item{target}{Target rotation matrix. Usually obtained with \cite{make_target()}.}

\item{bifactor}{Logical. Set to TRUE if model is bifactor.}

\item{fm}{Factor extraction method. Defaults to "Principal Axis", see \cite{psych::fa()}
for alternatives.}

\item{rotate}{Character vector. Factor rotation to use.
Defaults to oblique target rotation ("targetQ"). Automatically set to
orthogonal target rotation if \code{bifactor =TRUE}.}

\item{...}{Aditional parameters passed to \cite{psych::fa()}.}
}
\value{
\cite{psych::fa()} object with factor extraction results.
}
\description{
Wrapper around \cite{psych::fa()} for running ESEM-like exploratory factor analysis.
}
